launch event activity on user auth
