+++ app/src/main/java/com/microsoft/office/EventRobot/MainActivity.java
@@ -58,68 +58,5 @@ public class MainActivity extends AppCompatActivity {
 
         return super.onOptionsItemSelected(item);
     }
-    @Override
-    public void onProvideAssistContent(AssistContent assistContent) {
-        //  super.onProvideAssistContent(assistContent);
 
-        String structuredJson = null;
-        try {
-            String attendeeJson = new JSONObject()
-                    .put("@context","http://schema.org")
-                    .put("@type","person")
-                    .put("familyName","Smith")
-                    .put("givenName","Brad").toString();
-            String postalAddressJson = new JSONObject()
-                    .put("@context","http://schema.org")
-                    .put("@type","postalAddress")
-                    .put("postalCode","98466")
-                    .put("streetAddress","9726 Vision court")
-                    .put("addressLocality","Garden Grove")
-                    .put("addressRegion","CA").toString();
-
-            structuredJson = new JSONObject()
-                    .put("@context","http://schema.org")
-                    .put("@type","event")
-                    .put("attendee", attendeeJson)
-                    .put("location", postalAddressJson)
-                    .put("startDate", "2017-03-06T19:30:00-08:00")
-                    .toString();
-
-//            structuredJson = "{\n" +
-//                    "  \"@context\": \"http://schema.org\",\n" +
-//                    "  \"@type\": \"EventReservation\",\n" +
-//                    "  \"reservationNumber\": \"E123456789\",\n" +
-//                    "  \"reservationStatus\": \"http://schema.org/Confirmed\",\n" +
-//                    "  \"underName\": {\n" +
-//                    "    \"@type\": \"Person\",\n" +
-//                    "    \"name\": \"John Smith\"\n" +
-//                    "  },\n" +
-//                    "  \"reservationFor\": {\n" +
-//                    "    \"@type\": \"Event\",\n" +
-//                    "    \"name\": \"Foo Fighters Concert\",\n" +
-//                    "    \"startDate\": \"2017-03-06T19:30:00-08:00\",\n" +
-//                    "    \"location\": {\n" +
-//                    "      \"@type\": \"Place\",\n" +
-//                    "      \"name\": \"AT&T Park\",\n" +
-//                    "      \"address\": {\n" +
-//                    "        \"@type\": \"PostalAddress\",\n" +
-//                    "        \"streetAddress\": \"24 Willie Mays Plaza\",\n" +
-//                    "        \"addressLocality\": \"San Francisco\",\n" +
-//                    "        \"addressRegion\": \"CA\",\n" +
-//                    "        \"postalCode\": \"94107\",\n" +
-//                    "        \"addressCountry\": \"US\"\n" +
-//                    "      }\n" +
-//                    "    }\n" +
-//                    "  }\n" +
-//                    "}";
-
-            Log.i("Json", structuredJson);
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-
-        assistContent.setStructuredData(structuredJson);
-        super.onProvideAssistContent(assistContent);
-
-    }
 }
